backend/app/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
gunicorn
pandas>=2.1.0
numpy>=1.26.0
google-cloud-bigquery>=3.9.0
db-dtypes>=1.2.0
pyarrow>=12.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
google-cloud-storage==3.8.0
